Potentially fixed experience level limitation in the relic research UI

+++ src/main/java/it/hurts/sskirillss/relics/client/screen/description/ability/widgets/RerollAbilityActionWidget.java
@@ -96,7 +96,7 @@ public class RerollAbilityActionWidget extends AbstractAbilityActionWidget {
         int renderWidth = 0;
 
         int requiredExperience = relic.getRerollPlayerExperienceCost(getScreen().getStack(), getAbility());
-        int experience = EntityUtils.getPlayerTotalExperience(minecraft.player);
+        long experience = EntityUtils.getPlayerTotalExperience(minecraft.player);
 
         MutableComponent negativeStatus = Component.translatable("tooltip.relics.relic.status.negative");
         MutableComponent positiveStatus = Component.translatable("tooltip.relics.relic.status.positive");
